requirements.txt
langchain==0.0.294
openai==0.28.0
python-dotenv==1.0.0
web3==6.9.0